backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
gunicorn==23.0.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==5.1.5
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
kafka-python-ng==2.2.3
kombu==5.4.2
mock==5.1.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pytest==8.3.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
